apis/auth: cleanup dependency functions
