+++ byoeb-v1/byoeb/byoeb/apis/chat.py
@@ -16,7 +16,7 @@ from byoeb.models.message_category import MessageCategory
 from byoeb.services.user.utils import get_user_ids_from_phone_number_ids
 from fastapi import APIRouter, Query, Body, Depends
 from fastapi.responses import JSONResponse
-from byoeb.apis.auth import verify_whatsapp_signature, get_active_phone_id, get_current_user, require_permissions, require_tenant
+from byoeb.services.auth.dependencies import get_active_phone_id, get_current_user, require_permissions, require_tenant, verify_whatsapp_signature
 from byoeb.services.auth.models import AuthPermission
 import byoeb.services.chat.constants as chat_constants
 import byoeb.utils.utils as utils
